requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
asynctest==0.13.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
Flask==2.2.2
frozenlist==1.3.3
gevent==22.10.2
greenlet==2.0.2
grequests==0.6.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
pycparser==2.21
pymongo==4.3.3
python-bsonjs==0.3.0
requests==2.28.2
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.12.1
zope.event==4.6
zope.interface==5.5.2
